Move counter into pkg and update imports
Relocate counter.py into swanlab/sdk/internal/core_python/pkg/counter/__init__.py and update package exports and imports accordingly. Add counter to pkg.__init__.py and update imports in core_python/__init__.py and record_builder.py to import Counter from swanlab.sdk.internal.core_python.pkg.counter. This organizes the counter into the pkg namespace and keeps imports consistent.
