+++ swanlab/sdk/internal/core_python/__init__.py
@@ -41,8 +41,8 @@ from swanlab.sdk.internal.core_python.api.experiment import (
     stop_experiment,
 )
 from swanlab.sdk.internal.core_python.api.project import get_or_create_project, get_project
-from swanlab.sdk.internal.core_python.counter import Counter
 from swanlab.sdk.internal.core_python.metrics import RunMetrics
+from swanlab.sdk.internal.core_python.pkg.counter import Counter
 from swanlab.sdk.internal.core_python.store import DataStoreWriter
 from swanlab.sdk.internal.core_python.transport import Transport
 from swanlab.sdk.internal.core_python.transport.sender import HttpRecordSender
